chore: utoo-web code styles
